refactor: extract <TextInput> logic
